requirements.txt
aiogram
aiohttp==3.7.4
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
idna==3.1
multidict==5.1.0
python-dotenv==0.15.0
pytz==2021.1
typing-extensions==3.7.4.3
yarl==1.6.3
oauth2>=1.9.0.post1
oauthlib>=1.0.3
requests>=2.8.1
requests-oauthlib>=0.5.0
